Updated variable type for addPost to int. Able to delete posts.

+++ src/main/java/com/example/springbootspillit/controller/PostController.java
@@ -25,7 +25,7 @@ public class PostController {
 
 
     @DeleteMapping("/{postId}")
-    public void deletePostById(@PathVariable Long postId){
+    public void deletePostById(@PathVariable int postId){
         postService.deleteById(postId);
     }
